requirements.txt
ib_insync>=0.9.86
PyYAML>=6.0
